v 0.1.2 Formatting and changes for export
-Added an int value to all rows to ensure the export part produces files
in the same order (Currently the same as English, might change it to
keep the order of other languages)

-Also keeping blank rows for the same purpose

-Both the column with the values and the blank rows are hidden.

Export is still not ready, hopefully tonight/tomorrow.
Hiding seems to not work on OO, which is annoying, will figure it out
later.
Will prob unhide the blank rows and switch the order values to formulas
so that it can all be edited in worksheet and then exported to .lang
without further changes.